fix: replaced double for decimal for screen size

+++ src/Catalog/Catalog.Application/DTOs/Filtering/Components/DisplayFilter.cs
@@ -5,8 +5,8 @@ namespace Catalog.Application.DTOs.Filtering.Components
     public record DisplayFilter : ComponentFilter
     {
         public ProductType? DesignedFor { get; set; }
-        public double? MinScreenSizeInches { get; set; }
-        public double? MaxScreenSizeInches { get; set; }
+        public decimal? MinScreenSizeInches { get; set; }
+        public decimal? MaxScreenSizeInches { get; set; }
         public string? Resolution { get; set; }
         public int? MinRefreshRateHz { get; set; }
         public int? MaxRefreshRateHz { get; set; }
